chore: rename legacy logo assets to onenew
