+++ frontend/src/media/logo/onenew-icon.png
