+++ frontend/src/pages/GeneralSettings/EmbeddingPreference/index.jsx
@@ -3,7 +3,7 @@ import Sidebar from "@/components/SettingsSidebar";
 import { isMobile } from "react-device-detect";
 import System from "@/models/system";
 import showToast from "@/utils/toast";
-import OneNewIcon from "@/media/logo/anything-llm-icon.png";
+import OneNewIcon from "@/media/logo/onenew-icon.png";
 import OpenAiLogo from "@/media/llmprovider/openai.png";
 import AzureOpenAiLogo from "@/media/llmprovider/azure.png";
 import GeminiAiLogo from "@/media/llmprovider/gemini.png";
