+++ frontend/src/pages/OnboardingFlow/Steps/DataHandling/index.jsx
@@ -1,6 +1,6 @@
 import PreLoader from "@/components/Preloader";
 import System from "@/models/system";
-import OneNewIcon from "@/media/logo/anything-llm-icon.png";
+import OneNewIcon from "@/media/logo/onenew-icon.png";
 import OpenAiLogo from "@/media/llmprovider/openai.png";
 import GenericOpenAiLogo from "@/media/llmprovider/generic-openai.png";
 import AzureOpenAiLogo from "@/media/llmprovider/azure.png";
