+++ frontend/src/pages/OnboardingFlow/Steps/Home/index.jsx
@@ -3,7 +3,7 @@ import LGroupImg from "./l_group.png";
 import RGroupImg from "./r_group.png";
 import LGroupImgLight from "./l_group-light.png";
 import RGroupImgLight from "./r_group-light.png";
-import OneNewLogo from "@/media/logo/anything-llm.png";
+import OneNewLogo from "@/media/logo/onenew.png";
 import { useNavigate } from "react-router-dom";
 import { useTheme } from "@/hooks/useTheme";
 import { useTranslation } from "react-i18next";
