+++ server/.gitignore
@@ -2,7 +2,8 @@
 .env.development
 .env.test
 storage/assets/*
-!storage/assets/anything-llm.png
+!storage/assets/onenew.png
+!storage/assets/onenew-dark.png
 storage/documents/*
 storage/comkey/*
 storage/tmp/*
