+++ server/prisma/seed.js
@@ -4,7 +4,7 @@ const prisma = new PrismaClient();
 async function main() {
   const settings = [
     { label: "multi_user_mode", value: "false" },
-    { label: "logo_filename", value: "anything-llm.png" },
+    { label: "logo_filename", value: "onenew.png" },
   ];
 
   for (let setting of settings) {
