+++ server/swagger/index.js
@@ -21,8 +21,8 @@ function waitForElm(selector) {
 // Force change the Swagger logo in the header
 waitForElm('.topbar-wrapper').then((elm) => {
   if (window.SWAGGER_DOCS_ENV === 'development') {
-    elm.innerHTML = `<img href='${window.location.origin}' src='http://localhost:3000/public/anything-llm-light.png' width='200'/>`
+    elm.innerHTML = `<img href='${window.location.origin}' src='http://localhost:3000/public/onenew.png' width='200'/>`
   } else {
-    elm.innerHTML = `<img href='${window.location.origin}' src='${window.location.origin}/anything-llm-light.png' width='200'/>`
+    elm.innerHTML = `<img href='${window.location.origin}' src='${window.location.origin}/onenew.png' width='200'/>`
   }
 });
